refactor: reorganize span filter
