+++ src/phoenix/trace/dsl/__init__.py
@@ -1,0 +1,5 @@
+from phoenix.trace.dsl.filter import SpanFilter
+
+__all__ = [
+    "SpanFilter",
+]
